change to singleton
